Rename ServiceHandler and TunnelService to ProcessHandler and TunnelProcess
